Created a very basic loading system that prevent scene loading until finished
Created a basic loading system that will display the a loading bar and the progress of loading asset, data, scene from any source with a text description.
The Loading system will prevent loading next scene until loading is complete. UI Panel loading is still possible though.

switch from Photon to Mirror because it is opensource an viable solution for networking for gacha game that reduces complication.