Built on Network Manager
Built on Network Manager.
It now Open the Server Console by itself.
Created a seperate task to detect if the player is connect to wifi which will check every 2 second and recieve a callback when the bool value changes.
Minor changes with the server.

Known issue
Task doesn't get disposed of when application quit.
Server process doesn't close because IHost isn't disposed of within either the Server code or NetworkManager.cs

Next Task
Got to format and Modularize the network script and server script.
Got to link the game logic with the network logic