Created in-memory database + message packer support for query loading. Cleaned full solution, optimized,refactor, organized folder structure
Created support for in-memory database with messagepack support to contain data in database. currently applied it for skill which saves Id, name, description and a byte[] for texture data.

Data get saved as bytes

1 texture byte[] with name saving is equal to 27 kb.  1000 skill data is approx 27 mb of data

Optimized and refactor the full solution for better readability, preformance, and to work with the in-memory database (continuing)

organized folder structure