Continue building on in-memory database
Created more tables to store in in-memory database (Ability, Character, Equipment, Item, Material, Skill, Weapon). Still not finished creating the tables to store to database. creating in-memory database handle to prevent it from reloading bytes into file if the file has been created and the bytes are the same(continuing). Completely remove scriptable object Addressable database implementation for in-memory database.