Kinematica 3 character Idle animation passed
Created the bare bone or the idle animations for the 3 selectable character in the creation scene. The Animation Provider Is currently using the burst compiler, but it isn't doing alot . Going to add on to the Job for the animation overtime. Also Created a test Marker for kinematica which is basically an event that utilizes the burst compiler so no manged object can be in either the Job or Marker.