Getting started with ECS and still planning on the ECS/DOTS architecture
Created a dynamic add and remove and other mutable query that can happen on the in-memory database.

Abstracted the in-memory database class to hide any low-level code away. I have also abstracted away the global setting script that handles all the file path to load and read from.

next task will be creating the ECS/DOTS architecture and make it work with the in-memory database architecture and the UI system in a clean way.

The in-memory database is still not complete but i will work on it over time.