Hooked up the UI System with the In-MemoryDatabase.
I've hooked up the in-memory database with the ui system by doing in-memory database query in the UI system.

Also the in-memory database can do dynamic appending, removing and replacing. So each new unlocked character will append it data to the in-memory database and it will be cached as a binary file so read and doing query are very quick.

currently the hook up between the UI system and the in-memory database is quite primitive, can be optimized, and refactored. Which will be done over the course of the development