Created a way for the UI system to query the in-memorydatabase and recieve data
Create a way for the UI System (UI Builder) to query the dabase depending on a primary or secondary value and return the recieved data from the database so the game can know which data was received an respone to it through a sub/pub way (subscribe/publish).

Going to work on hook between the animation system and the in-memorydatabase.