Created the basic hook between the UI system and the animation system
create a basic hook which will improve overtime between the animation system and the UI system using the in-memorydatabase and the messagebroker to communicate between both the UI and the animation depending on a database query.

Animation system now uses ComponentSystem instead of the snapshot provider. To utilize ECS/DOTS. So the animation system utilizes the following(DOTS/ECS, Jobs, in-memorydatabase). It also uses Ulid which is faster then GUID for in-memorydatabase comparision to achieve optimal performance. The animation system create an Job per update for the Kinematica which is utilizing the burst compiler.

Overall the systems will be improved overtime.