Optimized Currency system and created a stackable inventory system
Optimzed Currency system by removing the foreach loop when the wallet amount change and comparing the collection key with the IQuantifiable key. This get called every time you add or remove money from the wallet, so frequently. It also make it less readable and it will eat performance when more currency is added on to the collection or removed.
To fix it i replaced it with a ContainsKey to see if it contains the same key with the IQuantifiable key if so send a event otherwise don't. rather then iterating the collection. Also create GameItem as a stackable inventory system.

Also the GameInventory now only let you retrieve the currently add or removed item from the Observable subscription.

Tested. GC calls are small and some are calls are zero and ms count is close to zero.