Created Transaction System and tested
The transaction system has been created and tested it work with the wallet system, inventory system without know about each other.

Transaction will see if player has sufficient cost and will handle the virtual transaction and give the player the payout if it is a success and from there it will add/remove it to the inventory system and wallet system.

Transaction will be refactor as development continues and re tested