Created Store System and tested
Created store system and tested it.
Unit test and performance test.
The store system contains a collection of Transaction that has been added to the store thus having a reference to the gameTransaction system for purchasing a transaction.

gamestore check if the transaction exisit in the store and if so it will call the transaction.purchase on it to purchase the transaction.