Shop UI now work with game foundation
Shop UI now dynamically create button for each transaction in the given store. Each button represent a transaction and will start the transaction if the button is pressed.
 The transaction also check if the user has the correct amount of payment before giving the player the payout.
Created a callback event to be called when the transaction was successful and when it failed to enable me to show a UI pop-up of the transaction state to the user, which I will be doing later.
Once the transaction succeeds the reward  will automagically be added (currency/inventory);.