A small ways of creating the combat system in UnitTechTest Project
Created UnitQueueSystem to wait for a given amount of time before the player can attack and a UnitRaycastSystem which will store the target to the player once the UnitQueueSystem says the player has waited for the correct minimum duration.
both these ECS/DOTS scripts need UnitRuntime component, UnitExecutionRuntime, and EnemyRuntime.

Issues:
The UnitRaycastSystem can utilize Schedule due do the NativeQueue throwing a big exception and terminating unity. A solution is to upgrade collection package, but it will not be compatible with kinematica or just keep it as Run()