Very rough implementation of battle animation (currently on finished moving to target)
Created a very rough animation system using kinematica which work with ECS/DOTS. The scripts will be constantly be improved.
The Pipeline for playing the animation is (UnitQueueSystem -> UnitRaycastSystem -> UnitAnimation)

The animation system uses root motion and query the collection of audio clip for the clip that has the best trajectory to the target.

*NOTE rebuild any kinematica asset before testing or it will throw an exception.